Test help for main command
